#82 fix: CounselorService와 WishListCounselorService 순환참조 문제 해결
